Made Enemy read from XML
Enemy now uses XMLReader to read an XML file containing the amount of
enemies to create for a wave.

Still need to find a way to spawn a new wave after the current one is
gone. But as soon as that is figured out the enemy creation code should
still be completely compatible, since it just checks the states of the
current wave and level and pulls the right info from the XML.