Refactored to OSGi
Had to make a new project for this (hence the removal of the old
project)

DynUpdater actually adds plugins in runtime!! WOOO
Still working on removing plugins during runtime though.

Need to figure out what the right path to assets are, because they are
NOT loading right now, which is like the only major issue left at this
point.